requirements.txt
langchain
langchain-community
streamlit
openai
tiktoken
python-dotenv
pinecone-client
duckduckgo-search